Commit2

+++ app/src/main/java/com/example/itogovoe/data/api/CurrencyResponse.kt
@@ -5,5 +5,5 @@ data class CurrencyResponse(
     val timestamp: Long,
     val base: String?,
     val date: String,
-    val rates: Map<String, Double>
+    val rates: Map<String, Float>
 )
